requirements.txt
# 语音交互
RealtimeSTT>=0.3.0
openai-whisper>=20231117
TTS>=0.22.0

# 意图理解
requests>=2.31.0  # 用于调用 Ollama API

# 邮件处理
IMAPClient>=3.0.0
exchangelib>=5.0.0

# 文件操作
watchdog>=3.0.0

# 系统集成
pystray>=0.19.0
pynput>=1.7.0
Pillow>=10.0.0

# 安全
cryptography>=41.0.0

# 配置和日志
python-dotenv>=1.0.0
